refactor: move Table component to lib/components
